ManagedArray now uses both chai arrays and std::vectors

+++ src/components/core/src/managers/Functions/CompositeFunction.cpp
@@ -112,7 +112,7 @@ void CompositeFunction::InitializeFunction()
 {
   // Register variables
   string_array const & variables = getReference<string_array>(keys::variableNames);
-  for (int ii=0 ; ii<variables.size() ; ++ii)
+  for (int ii=0 ; ii<static_cast<int>(variables.size()) ; ++ii)
   {
     parserContext.addVariable(variables[ii].c_str(), ii * sizeof(double));
   }
